+++ scripts/tmux_scrollback_clear.sh
@@ -1,0 +1,11 @@
+#!/usr/bin/env bash
+
+CURRENT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
+source $CURRENT_DIR/shared.sh
+
+main() {
+	tmux clear-history
+	display_message "Pane history cleared!"
+}
+main