+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### master
 - add other plugins list to the README
 - update readme to reflect github organization change
+- add a 'clear pane history' key binding
 
 ### v0.0.1, 2014-06-02